# src/main/java/com/example/bankcards/dto/UserResponse.java
package com.example.bankcards.dto;

import lombok.Data;
import java.time.LocalDateTime;

@Data
public class UserResponse {
    private Long id;
    private String username;
    private String email;
    private String role;
    private Boolean enabled;
    private LocalDateTime createdAt;
    private Long cardsCount;
}